init
boundaries are not yet implemented

+++ src/constants.rs
@@ -1,5 +1,5 @@
-pub const SIMULATION_WIDTH: u32 = 500;
-pub const SIMULATION_HEIGHT: u32 = 500;
+pub const SIMULATION_WIDTH: usize = 500;
+pub const SIMULATION_HEIGHT: usize = 500;
 pub const PIXEL_SIZE: u32 = 1;
 pub const NUM_INDEX: u32 = 9; //cur_bottom cur_left cur_top cur_right next_bottom next_left next_top next_right pressure
 pub const WALL_FAC: f32 = 1.;
